+++ csrc/secure_channels/sopc_secure_connection_state_mgr.h
@@ -25,6 +25,6 @@
 void SOPC_SecureConnectionStateMgr_Dispatcher(SOPC_SecureChannels_InputEvent event,
                                               uint32_t eltId,
                                               void* params,
-                                              int32_t auxParam);
+                                              uintptr_t auxParam);
 
 #endif /* SOPC_SECURE_CONNECTION_STATE_MGR_H_ */
